+++ stdnet/lib/redis/__init__.py
@@ -1,1 +1,0 @@
-from .redis import *