+++ CHANGELOG.rst
@@ -18,7 +18,7 @@ Ver. 0.6.0 - Development
 * Moved testing functions into the :mod:`stdnet.test` module.
 * Reorganized and expanded documentation.
 * Bug fix in :class:`stdnet.orm.PickleObjectField` field.
-* **254 regression tests** with **78%** coverage.
+* **261 regression tests** with **78%** coverage.
 
 Ver. 0.5.5 - 2011 June 6
 ============================